+++ google2/src/main/scala/org/broadinstitute/dsde/workbench/google2/GoogleBigQueryService.scala
@@ -4,14 +4,22 @@ import cats.effect.{Blocker, ContextShift, Resource, Sync, Timer}
 import com.google.auth.Credentials
 import com.google.cloud.ServiceOptions.getDefaultProjectId
 import com.google.cloud.bigquery.BigQueryOptions.DefaultBigQueryFactory
-import com.google.cloud.bigquery.{BigQuery, BigQueryOptions, JobId, QueryJobConfiguration, TableResult}
-import org.typelevel.log4cats.StructuredLogger
+import com.google.cloud.bigquery.{Acl, BigQuery, BigQueryOptions, DatasetId, JobId, QueryJobConfiguration, TableResult}
+import io.chrisdavenport.log4cats.StructuredLogger
+import org.broadinstitute.dsde.workbench.model.WorkbenchEmail
 import org.broadinstitute.dsde.workbench.model.google.GoogleProject
 
 trait GoogleBigQueryService[F[_]] {
   def query(queryJobConfiguration: QueryJobConfiguration, options: BigQuery.JobOption*): F[TableResult]
 
   def query(queryJobConfiguration: QueryJobConfiguration, jobId: JobId, options: BigQuery.JobOption*): F[TableResult]
+
+  def createDataset(datasetName: String,
+                    labels: Map[String, String],
+                    aclBindings: Map[Acl.Role, Seq[(WorkbenchEmail, Acl.Entity.Type)]]
+  ): F[DatasetId]
+
+  def deleteDataset(datasetName: String): F[Boolean]
 }
 
 object GoogleBigQueryService {
@@ -20,6 +28,13 @@ object GoogleBigQueryService {
     blocker: Blocker
   ): Resource[F, GoogleBigQueryService[F]] =
     credentialResource(pathToJson) flatMap (resource(_, blocker))
+
+  def resource[F[_]: Sync: ContextShift: Timer: StructuredLogger](
+    pathToJson: String,
+    projectId: GoogleProject,
+    blocker: Blocker
+  ): Resource[F, GoogleBigQueryService[F]] =
+    credentialResource(pathToJson) flatMap (resource(_, blocker, projectId))
 
   def resource[F[_]: Sync: ContextShift: Timer: StructuredLogger](
     credentials: Credentials,